API

IRMA.SHist

IRMA.Stopwatch

IRMA.Stopwatch

IRMA.Stopwatch

OnlineStats.Hist

IRMA.asNamedTuple

IRMA.deserializeArray

IRMA.mergeStatsCollectionWithSHist

IRMA.mpiGatherSerialized

IRMA.partitionDS

IRMA.rankTimings

IRMA.rankTotalTime

IRMA.stamp

IRMA.SHist — Method

SHist(h::Hist)

Create a SHist (Static Histogram) from an already filled Hist. 

Note that the SHist is immutable.

IRMA.Stopwatch — Type

Stopwatch

Stopwatch is an object that keeps track of MPI.Wtime when asked.

At construction time, the "start" time is recorded.
Call `stamp(sw, stamp)` to record the current MPI.Wtime and label
                        it with "stamp"

Call `asNamedTuple(sw)` for transferring to other MPI ranks. The
resulting NamedTuple with be `isbits` type.

IRMA.Stopwatch — Method

Stopwatch(nt::NamedTuple)

Create a Stopwatch from a previously filled named tuple

IRMA.Stopwatch — Method

Stopwatch()

Create a Stopwatch. The "start" entry will automatically be made
and the MPI.Wtime will be filled in.

OnlineStats.Hist — Method

Hist(sh::SHist)

Create an OnlineStats.Hist from a SHist

IRMA.asNamedTuple — Method

asNamedTuple(sw::Stopwatch)

Convert `sw` into a `NamedTuple` for MPI transport.

IRMA.deserializeArray — Method

deserializeArray(a, s) When you use MPI.Gather, you get one long array with all of the contents from the ranks mushed together. They need to be separated and deserialized.

Returns an array of deserialized objects

a is the array of data all mushed together
s is an array of the data size for each rank

IRMA.mergeStatsCollectionWithSHist — Method

mergeStatsCollectionWithSHist(s1::Series, s2::Series)

Because Static Histograms are immutable, we cannot use the standard `OnlineStats.merge` 
    function (actually, it is `OnlineStatsBase.merge`) because the underlying function 
    is `merge!`.

IRMA.mpiGatherSerialized — Method

mpiGatherSerialized(obj, isroot, root, comm)

Serializes the object, determines the size, calls MPI.Allgather on the sizes,
calls MPI.Allgather on the serialized data, deserializes the data.

Returns an array of deserialized data from all of the ranks

obj is the object to serialize and send
isroot is a boolean which is true if this rank is the root rank
root is the root rank id
comm is the MPI communicator

IRMA.partitionDS — Method

partitionDS(dsLength, nRanks)

Given the length of a dataset (or anything, really), determine and return partitions over 
nRanks MPI ranks that are as close to the same size as possible. This is really just a wrapper 
around Distributed.splitrange with some added error checking to produce nice messages.

IRMA.rankTimings — Method

rankTimings(arrayOfNamedTuples)

Process an array of named tuples (e.g. from MPI ranks from `asNamedTuple`
that was gathered and saved) turning them into an array of NamedTuples
of timing differences for each step.

IRMA.rankTotalTime — Method

rankTotalTime(arrayOfNamedTuples)

Return an array for the total time of each ranks

IRMA.stamp — Method

stamp(sw::Stopwatch, stamp::String)

For a `Stopwatch` `sw`, record the MPI time and the stamp.

It returns the elapsed time from the previous stamp.