API

IRMA.SHist

OnlineStats.Hist

IRMA.mergeStatsCollectionWithSHist

IRMA.partitionDS

IRMA.mergeStatsCollectionWithSHist — Method

mergeStatsCollectionWithSHist(s1::Series, s2::Series)

Because Static Histograms are immutable, we cannot use the standard `OnlineStats.merge` 
    function (actually, it is `OnlineStatsBase.merge`) because the underlying function 
    is `merge!`.

IRMA.partitionDS — Method

partitionDS(dsLength, nRanks)

Given the length of a dataset (or anything, really), determine and return partitions over 
nRanks MPI ranks that are as close to the same size as possible. This is really just a wrapper 
around Distributed.splitrange with some added error checking to produce nice messages.

IRMA.SHist — Method

SHist(h::Hist)

Create a SHist (Static Histogram) from an already filled Hist. 

Note that the SHist is immutable.

OnlineStats.Hist — Method

Hist(sh::SHist)

Create an OnlineStats.Hist from a SHist